Add files via upload
Generates commands for drawing several letters such as A B T or Z.  Does not adjust for new lines at the moment.  Is very dynamic at easily adjusting for a different width/height draw area as well as spacing between characters.  Adjustment for which planes are where is also an easy fix.